requirements.txt
caldav
python-dotenv
mcp
mcp[cli]